Refactor functions for windows dataplane and fix winfv
