+++ go.mod
@@ -21,7 +21,7 @@ require (
 	github.com/nmrshll/go-cp v0.0.0-20180115193924-61436d3b7cfa
 	github.com/onsi/ginkgo v1.10.1
 	github.com/onsi/gomega v1.7.0
-	github.com/projectcalico/libcalico-go v1.7.2-0.20200728221526-a5e781268cd1
+	github.com/projectcalico/libcalico-go v1.7.2-0.20200807225946-83627ff7d609
 	github.com/prometheus/common v0.4.1
 	github.com/rakelkar/gonetsh v0.0.0-20190930180311-e5c5ffe4bdf0
 	github.com/satori/go.uuid v1.2.0
@@ -42,5 +42,3 @@ require (
 	k8s.io/client-go v0.17.3
 	k8s.io/utils v0.0.0-20191114200735-6ca3b61696b6
 )
-
-replace github.com/projectcalico/libcalico-go => github.com/song-jiang/libcalico-go v1.6.1-0.20200805171439-12372927ea4c
