+++ pkg/dataplane/grpc/grpc_dataplane.go
@@ -150,24 +150,3 @@ func (d *grpcDataplane) CleanUpNamespace(args *skel.CmdArgs) error {
 	}
 	return nil
 }
-
-// This is a dummy function required for Windows
-func (d *grpcDataplane) NetworkApplicationContainer(args *skel.CmdArgs) error {
-	return nil
-}
-
-func (d *grpcDataplane) EnsureVXLANTunnelAddr(ctx context.Context, calicoClient calicoclient.Interface, nodeName string, ipNet *net.IPNet) error {
-	return nil // No-op on Linux.
-}
-
-func (d *grpcDataplane) MaintainWepDeletionTimestamps(timeout int) error {
-	return nil
-}
-
-func (d *grpcDataplane) CheckWepJustDeleted(containerID string, timeout int) (bool, error) {
-	return false, nil
-}
-
-func (d *grpcDataplane) RegisterDeletedWep(containerID string) error {
-	return nil
-}
