+++ win_tests/calico_cni_k8s_windows_test.go
@@ -67,6 +67,19 @@ func deleteNamespace(clientset *kubernetes.Clientset, name string) {
 	}
 }
 
+func updateIPAMStrictAffinity(calicoClient client.Interface, enabled bool) {
+	// Currently only Linux host is able to update IPAMConfig.
+	// Use os override to fake a linux host.
+	ctx := context.WithValue(context.Background(), "windowsHost", "linux")
+	ipamConfig, err := calicoClient.IPAM().GetIPAMConfig(ctx)
+	Expect(err).NotTo(HaveOccurred())
+
+	ipamConfig.StrictAffinity = enabled
+
+	err = calicoClient.IPAM().SetIPAMConfig(ctx, *ipamConfig)
+	Expect(err).NotTo(HaveOccurred())
+}
+
 var _ = Describe("Kubernetes CNI tests", func() {
 	var hostname string
 	networkName := "calico-fv"
@@ -109,6 +122,9 @@ var _ = Describe("Kubernetes CNI tests", func() {
 			caliNode, err := calicoClient.Nodes().Create(context.Background(), caliNode, options.SetOptions{})
 			Expect(err).NotTo(HaveOccurred(), "Failed to create Calico Node resource")
 		}
+
+		// Force StrictAffinity to be true, otherwise no IP allocation is possible.
+		updateIPAMStrictAffinity(calicoClient, true)
 	})
 
 	utils.ConfigureLogging("info")
@@ -1881,8 +1897,6 @@ var _ = Describe("Kubernetes CNI tests", func() {
 			"Namespace": testutils.HnsNoneNs,
 		})
 
-		dataplane := windows.NewWindowsDataplane(conf, logger)
-
 		clientset, err = k8s.NewK8sClient(conf, logger)
 		if err != nil {
 			panic(err)
@@ -1945,14 +1959,14 @@ var _ = Describe("Kubernetes CNI tests", func() {
 
 		ensureTimestamp := func(id string) {
 			log.Infof("Ensure timestamp for pod deletion")
-			t, err := testutils.GetTimestampValue(windows.PodDeletedKey, id)
+			t, err := testutils.GetTimestampValue(utils.PodDeletedKey, id)
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(t.Before(time.Now())).To(Equal(true))
 		}
 
 		It("should create pod deletion timestamp in registry", func() {
 			log.Infof("Delete pod deletion subkey")
-			err = testutils.DeleteSubKey(windows.CalicoRegistryKey, windows.PodDeletedKeyString)
+			err = testutils.DeleteSubKey(utils.CalicoRegistryKey, utils.PodDeletedKeyString)
 			Expect(err).NotTo(HaveOccurred())
 
 			log.Infof("Creating container")
@@ -1967,7 +1981,7 @@ var _ = Describe("Kubernetes CNI tests", func() {
 				ensureTimestamp(containerID)
 			}()
 
-			keyExists, err := testutils.CheckRegistryKeyExists(windows.PodDeletedKey)
+			keyExists, err := testutils.CheckRegistryKeyExists(utils.PodDeletedKey)
 			Expect(err).NotTo(HaveOccurred())
 			Expect(keyExists).To(Equal(true))
 		})
@@ -1994,21 +2008,21 @@ var _ = Describe("Kubernetes CNI tests", func() {
 			Expect(err).ShouldNot(HaveOccurred())
 			ensureTimestamp(containerID2)
 
-			log.Infof("Sleeping 7 seonds")
+			log.Infof("Sleeping 7 seconds")
 			time.Sleep(time.Second * 7)
 
 			log.Infof("Checking timestamp container 1 %s", containerID1)
-			justDeleted, err := dataplane.CheckWepJustDeleted(containerID1, 12)
+			justDeleted, err := utils.CheckWepJustDeleted(containerID1, 12)
 			Expect(err).ShouldNot(HaveOccurred())
-			Expect(justDeleted).To(Equal(true))
+			//JustDeleted for container1 could vary because it depends on how long it takes to delete container2.
 
 			log.Infof("Checking timestamp container 2 %s", containerID2)
-			justDeleted, err = dataplane.CheckWepJustDeleted(containerID2, 12)
+			justDeleted, err = utils.CheckWepJustDeleted(containerID2, 12)
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(justDeleted).To(Equal(true))
 
 			log.Infof("Checking timestamp container 3 %s", containerID3)
-			justDeleted, err = dataplane.CheckWepJustDeleted(containerID3, 12)
+			justDeleted, err = utils.CheckWepJustDeleted(containerID3, 12)
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(justDeleted).To(Equal(false))
 
@@ -2020,15 +2034,15 @@ var _ = Describe("Kubernetes CNI tests", func() {
 			// Make sure timeout on pod1, pod2 deletion timestamp. 7+7 > 12
 			log.Infof("Sleeping further 7 seonds")
 			time.Sleep(time.Second * 7)
-			justDeleted, err = dataplane.CheckWepJustDeleted(containerID1, 12)
+			justDeleted, err = utils.CheckWepJustDeleted(containerID1, 12)
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(justDeleted).To(Equal(false))
 
-			justDeleted, err = dataplane.CheckWepJustDeleted(containerID2, 12)
+			justDeleted, err = utils.CheckWepJustDeleted(containerID2, 12)
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(justDeleted).To(Equal(false))
 
-			justDeleted, err = dataplane.CheckWepJustDeleted(containerID3, 12)
+			justDeleted, err = utils.CheckWepJustDeleted(containerID3, 12)
 			Expect(err).ShouldNot(HaveOccurred())
 			Expect(justDeleted).To(Equal(true))
 
@@ -2038,13 +2052,13 @@ var _ = Describe("Kubernetes CNI tests", func() {
 			Expect(err).Should(HaveOccurred())
 
 			// first two timestamp are gone
-			_, err = testutils.GetTimestampValue(windows.PodDeletedKey, containerID1)
+			_, err = testutils.GetTimestampValue(utils.PodDeletedKey, containerID1)
 			Expect(err).Should(HaveOccurred())
 
-			_, err = testutils.GetTimestampValue(windows.PodDeletedKey, containerID2)
+			_, err = testutils.GetTimestampValue(utils.PodDeletedKey, containerID2)
 			Expect(err).Should(HaveOccurred())
 
-			_, err = testutils.GetTimestampValue(windows.PodDeletedKey, containerID3)
+			_, err = testutils.GetTimestampValue(utils.PodDeletedKey, containerID3)
 			Expect(err).ShouldNot(HaveOccurred())
 		})
 	})
